LPS-60902 Move trash logic into new service
