+++ portal-impl/src/com/liferay/portlet/documentlibrary/service/impl/DLAppServiceImpl.java
@@ -55,8 +55,6 @@ import com.liferay.portal.service.ServiceContext;
 import com.liferay.portlet.documentlibrary.NoSuchFileEntryException;
 import com.liferay.portlet.documentlibrary.model.DLFolderConstants;
 import com.liferay.portlet.documentlibrary.service.base.DLAppServiceBaseImpl;
-import com.liferay.portlet.documentlibrary.service.permission.DLFileEntryPermission;
-import com.liferay.portlet.documentlibrary.service.permission.DLFileShortcutPermission;
 import com.liferay.portlet.documentlibrary.service.permission.DLFolderPermission;
 import com.liferay.portlet.documentlibrary.service.permission.DLPermission;
 import com.liferay.portlet.documentlibrary.util.DLAppUtil;
@@ -2114,108 +2112,6 @@ public class DLAppServiceImpl extends DLAppServiceBaseImpl {
 	}
 
 	/**
-	 * Moves the file entry from a trashed folder to the new folder.
-	 *
-	 * @param  fileEntryId the primary key of the file entry
-	 * @param  newFolderId the primary key of the new folder
-	 * @param  serviceContext the service context to be applied
-	 * @return the file entry
-	 */
-	@Override
-	public FileEntry moveFileEntryFromTrash(
-			long fileEntryId, long newFolderId, ServiceContext serviceContext)
-		throws PortalException {
-
-		Repository repository = repositoryProvider.getFileEntryRepository(
-			fileEntryId);
-
-		TrashCapability trashCapability = repository.getCapability(
-			TrashCapability.class);
-
-		FileEntry fileEntry = repository.getFileEntry(fileEntryId);
-
-		DLFileEntryPermission.check(
-			getPermissionChecker(), fileEntry, ActionKeys.UPDATE);
-
-		Folder destinationFolder = repository.getFolder(newFolderId);
-
-		return trashCapability.moveFileEntryFromTrash(
-			getUserId(), fileEntry, destinationFolder, serviceContext);
-	}
-
-	/**
-	 * Moves the file entry with the primary key to the trash portlet.
-	 *
-	 * @param  fileEntryId the primary key of the file entry
-	 * @return the file entry
-	 */
-	@Override
-	public FileEntry moveFileEntryToTrash(long fileEntryId)
-		throws PortalException {
-
-		Repository repository = repositoryProvider.getFileEntryRepository(
-			fileEntryId);
-
-		TrashCapability trashCapability = repository.getCapability(
-			TrashCapability.class);
-
-		FileEntry fileEntry = repository.getFileEntry(fileEntryId);
-
-		DLFileEntryPermission.check(
-			getPermissionChecker(), fileEntry, ActionKeys.DELETE);
-
-		return trashCapability.moveFileEntryToTrash(getUserId(), fileEntry);
-	}
-
-	/**
-	 * Moves the file shortcut from a trashed folder to the new folder.
-	 *
-	 * @param  fileShortcutId the primary key of the file shortcut
-	 * @param  newFolderId the primary key of the new folder
-	 * @param  serviceContext the service context to be applied
-	 * @return the file shortcut
-	 */
-	@Override
-	public FileShortcut moveFileShortcutFromTrash(
-			long fileShortcutId, long newFolderId,
-			ServiceContext serviceContext)
-		throws PortalException {
-
-		FileShortcut fileShortcut = getFileShortcut(fileShortcutId);
-
-		DLFileShortcutPermission.check(
-			getPermissionChecker(), fileShortcut, ActionKeys.UPDATE);
-
-		return dlAppHelperLocalService.moveFileShortcutFromTrash(
-			getUserId(), fileShortcut, newFolderId, serviceContext);
-	}
-
-	/**
-	 * Moves the file shortcut with the primary key to the trash portlet.
-	 *
-	 * @param  fileShortcutId the primary key of the file shortcut
-	 * @return the file shortcut
-	 */
-	@Override
-	public FileShortcut moveFileShortcutToTrash(long fileShortcutId)
-		throws PortalException {
-
-		Repository repository = repositoryProvider.getFileShortcutRepository(
-			fileShortcutId);
-
-		TrashCapability trashCapability = repository.getCapability(
-			TrashCapability.class);
-
-		FileShortcut fileShortcut = repository.getFileShortcut(fileShortcutId);
-
-		DLFileShortcutPermission.check(
-			getPermissionChecker(), fileShortcut, ActionKeys.DELETE);
-
-		return trashCapability.moveFileShortcutToTrash(
-			getUserId(), fileShortcut);
-	}
-
-	/**
 	 * Moves the folder to the new parent folder with the primary key.
 	 *
 	 * @param  folderId the primary key of the folder
@@ -2255,63 +2151,6 @@ public class DLAppServiceImpl extends DLAppServiceBaseImpl {
 		return moveFolders(
 			folderId, parentFolderId, fromRepository, toRepository,
 			serviceContext);
-	}
-
-	/**
-	 * Moves the folder with the primary key from the trash portlet to the new
-	 * parent folder with the primary key.
-	 *
-	 * @param  folderId the primary key of the folder
-	 * @param  parentFolderId the primary key of the new parent folder
-	 * @param  serviceContext the service context to be applied
-	 * @return the file entry
-	 */
-	@Override
-	public Folder moveFolderFromTrash(
-			long folderId, long parentFolderId, ServiceContext serviceContext)
-		throws PortalException {
-
-		Repository repository = repositoryProvider.getFolderRepository(
-			folderId);
-
-		TrashCapability trashCapability = repository.getCapability(
-			TrashCapability.class);
-
-		Folder folder = repository.getFolder(folderId);
-
-		DLFolderPermission.check(
-			getPermissionChecker(), folder, ActionKeys.UPDATE);
-
-		Folder destinationFolder = null;
-
-		if (parentFolderId != DLFolderConstants.DEFAULT_PARENT_FOLDER_ID) {
-			destinationFolder = repository.getFolder(parentFolderId);
-		}
-
-		return trashCapability.moveFolderFromTrash(
-			getUserId(), folder, destinationFolder, serviceContext);
-	}
-
-	/**
-	 * Moves the folder with the primary key to the trash portlet.
-	 *
-	 * @param  folderId the primary key of the folder
-	 * @return the file entry
-	 */
-	@Override
-	public Folder moveFolderToTrash(long folderId) throws PortalException {
-		Repository repository = repositoryProvider.getFolderRepository(
-			folderId);
-
-		TrashCapability trashCapability = repository.getCapability(
-			TrashCapability.class);
-
-		Folder folder = repository.getFolder(folderId);
-
-		DLFolderPermission.check(
-			getPermissionChecker(), folder, ActionKeys.DELETE);
-
-		return trashCapability.moveFolderToTrash(getUserId(), folder);
 	}
 
 	/**
@@ -2368,68 +2207,6 @@ public class DLAppServiceImpl extends DLAppServiceBaseImpl {
 
 		return repository.refreshFolderLock(
 			lockUuid, companyId, expirationTime);
-	}
-
-	/**
-	 * Restores the file entry with the primary key from the trash portlet.
-	 *
-	 * @param fileEntryId the primary key of the file entry
-	 */
-	@Override
-	public void restoreFileEntryFromTrash(long fileEntryId)
-		throws PortalException {
-
-		Repository repository = repositoryProvider.getFileEntryRepository(
-			fileEntryId);
-
-		TrashCapability trashCapability = repository.getCapability(
-			TrashCapability.class);
-
-		FileEntry fileEntry = repository.getFileEntry(fileEntryId);
-
-		DLFileEntryPermission.check(
-			getPermissionChecker(), fileEntry, ActionKeys.DELETE);
-
-		trashCapability.restoreFileEntryFromTrash(getUserId(), fileEntry);
-	}
-
-	/**
-	 * Restores the file shortcut with the primary key from the trash portlet.
-	 *
-	 * @param fileShortcutId the primary key of the file shortcut
-	 */
-	@Override
-	public void restoreFileShortcutFromTrash(long fileShortcutId)
-		throws PortalException {
-
-		FileShortcut fileShortcut = getFileShortcut(fileShortcutId);
-
-		DLFileShortcutPermission.check(
-			getPermissionChecker(), fileShortcut, ActionKeys.DELETE);
-
-		dlAppHelperLocalService.restoreFileShortcutFromTrash(
-			getUserId(), fileShortcut);
-	}
-
-	/**
-	 * Restores the folder with the primary key from the trash portlet.
-	 *
-	 * @param folderId the primary key of the folder
-	 */
-	@Override
-	public void restoreFolderFromTrash(long folderId) throws PortalException {
-		Repository repository = repositoryProvider.getFolderRepository(
-			folderId);
-
-		TrashCapability trashCapability = repository.getCapability(
-			TrashCapability.class);
-
-		Folder folder = repository.getFolder(folderId);
-
-		DLFolderPermission.check(
-			getPermissionChecker(), folder, ActionKeys.DELETE);
-
-		trashCapability.restoreFolderFromTrash(getUserId(), folder);
 	}
 
 	/**
